parser/requirements.txt
PyYAML==5.4.1
pymongo==3.12.0
requests==2.26.0
lxml==4.6.3